Test setup, few tests done, RegistrationController moved into SecurityController
