+++ .gitignore
@@ -15,3 +15,7 @@
 .phpunit.result.cache
 /phpunit.xml
 ###< symfony/phpunit-bridge ###
+
+###> code coverage
+TestCoverage
+###< code coverage
