+++ tests/Controller/DefaultControllerTest.php
@@ -1,0 +1,19 @@
+<?php
+
+
+namespace App\Tests\Controller;
+
+
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+
+class DefaultControllerTest extends WebTestCase
+{
+    public function testHomepage()
+    {
+        $client = static::createClient();
+
+        $client->request('GET', '/');
+
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+    }
+}
