+++ tests/Controller/UserControllerTest.php
@@ -1,0 +1,22 @@
+<?php
+
+
+namespace App\Tests\Controller;
+
+
+use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
+
+class UserControllerTest extends WebTestCase
+{
+    public function testAllRouteOnce()
+    {
+        $client = static::createClient();
+
+        $client->request('GET', '/users');
+
+        $this->assertEquals(200, $client->getResponse()->getStatusCode());
+
+        // todo :
+        // $client->request('GET', '/users/{id}/edit');
+    }
+}
